Added bert_tokenizer and overrode _featurize
